reorganize and clean contracts
